Update example to work with local environment
