+++ examples/src/authorization.ts
@@ -1,0 +1,15 @@
+import { zkdb } from './connection';
+
+// Check user existence then create
+if (!(await zkdb.auth.isUserExist('chiro-user'))) {
+  await zkdb.auth.signUp('chiro@orochi.network');
+}
+
+// Sign in
+await zkdb.auth.signIn();
+
+// Get information of current signed in user
+console.log('Authorization getUser: ', zkdb.auth.getUser());
+
+// Sign out
+await zkdb.auth.signOut();
