+++ examples/src/connection.ts
@@ -1,0 +1,13 @@
+import { ZkDatabase } from 'zkdb';
+
+// In reality you better to encrypt your private key and these information
+// It will be better if your load it from .env file
+export const zkdb = await ZkDatabase.connect({
+  userName: 'chiro-user',
+  privateKey: 'EKFTciRxyxshZjimay9sktsn7v5PvmC5zPq7q4JnitHUytxUVnFP',
+  environment: 'node',
+  // This URL is for local environment
+  url: 'http://0.0.0.0:4000/graphql',
+});
+
+export default zkdb;
