+++ examples/src/create-database.ts
@@ -1,14 +1,4 @@
-import { ZkDatabase } from 'zkdb';
-
-// In reality you better to encrypt your private key and these information
-// It will be better if your load it from .env file
-const zkdb = await ZkDatabase.connect({
-  userName: 'chiro-user',
-  privateKey: 'EKFTciRxyxshZjimay9sktsn7v5PvmC5zPq7q4JnitHUytxUVnFP',
-  environment: 'node',
-  // This URL is for test environment
-  url: 'https://test-serverless.zkdatabase.org/graphql',
-});
+import { zkdb } from './connection';
 
 // Check user existence then create
 if (!(await zkdb.auth.isUserExist('chiro-user'))) {
