+++ examples/src/create-document.ts
@@ -1,6 +1,6 @@
 import { Schema } from '@zkdb/common';
 import { CircuitString, UInt32 } from 'o1js';
-import { ZkDatabase } from 'zkdb';
+import { zkdb } from './connection';
 
 class Book extends Schema.create({
   name: CircuitString,
@@ -9,16 +9,6 @@ class Book extends Schema.create({
 }) {}
 
 export type TBook = typeof Book;
-
-// In reality you better to encrypt your private key and these information
-// It will be better if your load it from .env file
-const zkdb = await ZkDatabase.connect({
-  userName: 'chiro-user',
-  privateKey: 'EKFTciRxyxshZjimay9sktsn7v5PvmC5zPq7q4JnitHUytxUVnFP',
-  environment: 'node',
-  // This URL is for test environment
-  url: 'https://test-serverless.zkdatabase.org/graphql',
-});
 
 // Check user existence then create
 if (!(await zkdb.auth.isUserExist('chiro-user'))) {
