--- examples/src/features/create-collection.ts
@@ -1,64 +1,0 @@
-import { Schema } from '@zkdb/common';
-import { CircuitString, UInt32 } from 'o1js';
-import { Permission, ZkDatabase } from 'zkdb';
-
-class Book extends Schema.create({
-  name: CircuitString,
-  author: CircuitString,
-  release: UInt32,
-}) {}
-
-export type TBook = typeof Book;
-
-// In reality you better to encrypt your private key and these information
-// It will be better if your load it from .env file
-const zkdb = await ZkDatabase.connect({
-  userName: 'chiro-user',
-  privateKey: 'EKFTciRxyxshZjimay9sktsn7v5PvmC5zPq7q4JnitHUytxUVnFP',
-  environment: 'node',
-  // This URL is for test environment
-  url: 'https://test-serverless.zkdatabase.org/graphql',
-});
-
-// Check user existence then create
-if (!(await zkdb.auth.isUserExist('chiro-user'))) {
-  await zkdb.auth.signUp('chiro-user');
-}
-
-// Sign in
-await zkdb.auth.signIn();
-
-// Create new instance of `db_test`
-const dbTest = zkdb.db('db_test');
-
-// Create new instance of collection `db_test.book`
-const collectionBook = dbTest.collection<TBook>('book');
-
-// Define a group instance
-const librarian = dbTest.group('librarian');
-
-if (!(await librarian.exist())) {
-  await librarian.create({ groupDescription: 'Group of librarians' });
-}
-
-// Create an
-await collectionBook.create(
-  Book,
-  Permission.from({
-    owner: {
-      read: true,
-      write: true,
-      system: true,
-      delete: true,
-    },
-    // Other users in this group can read and write
-    group: {
-      read: true,
-      write: true,
-    },
-  }),
-  'librarian'
-);
-
-// Sign out
-await zkdb.auth.signOut();
